requirements.txt
python-telegram-bot==13.15
pycoingecko
requests
beautifulsoup4
